refactor : ParserTest, InputValidatorTest 디렉토리명 변경
- domain -> util로 테스트 디렉토리명 변경
