+++ src/test/java/calculator/util/InputValidatorTest.java
@@ -1,8 +1,8 @@
-package calculator.domain;
+package calculator.util;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import calculator.util.InputValidator;
+import calculator.domain.UserInput;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
