+++ src/test/java/calculator/util/ParserTest.java
@@ -1,8 +1,7 @@
-package calculator.domain;
+package calculator.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import calculator.util.Parser;
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
